GS/Vulkan: Combine GSDeviceVK and VKContext
Less indirection.
