+++ pcsx2/GS/Renderers/Vulkan/VKLoader.cpp
@@ -15,7 +15,8 @@
 
 #include "GS/Renderers/Vulkan/VKLoader.h"
 
-#include <atomic>
+#include "common/Assertions.h"
+
 #include <cstdarg>
 #include <cstdio>
 #include <cstdlib>
@@ -54,20 +55,19 @@ void Vulkan::ResetVulkanLibraryFunctionPointers()
 
 #if defined(_WIN32)
 
-static HMODULE vulkan_module;
-static std::atomic_int vulkan_module_ref_count = {0};
+static HMODULE s_vulkan_module;
+
+bool Vulkan::IsVulkanLibraryLoaded()
+{
+	return s_vulkan_module != NULL;
+}
 
 bool Vulkan::LoadVulkanLibrary()
 {
-	// Not thread safe if a second thread calls the loader whilst the first is still in-progress.
-	if (vulkan_module)
-	{
-		vulkan_module_ref_count++;
-		return true;
-	}
+	pxAssertRel(!s_vulkan_module, "Vulkan module is not loaded.");
 
-	vulkan_module = LoadLibraryA("vulkan-1.dll");
-	if (!vulkan_module)
+	s_vulkan_module = LoadLibraryA("vulkan-1.dll");
+	if (!s_vulkan_module)
 	{
 		std::fprintf(stderr, "Failed to load vulkan-1.dll\n");
 		return false;
@@ -75,7 +75,7 @@ bool Vulkan::LoadVulkanLibrary()
 
 	bool required_functions_missing = false;
 	auto LoadFunction = [&](FARPROC* func_ptr, const char* name, bool is_required) {
-		*func_ptr = GetProcAddress(vulkan_module, name);
+		*func_ptr = GetProcAddress(s_vulkan_module, name);
 		if (!(*func_ptr) && is_required)
 		{
 			std::fprintf(stderr, "Vulkan: Failed to load required module function %s\n", name);
@@ -90,45 +90,41 @@ bool Vulkan::LoadVulkanLibrary()
 	if (required_functions_missing)
 	{
 		ResetVulkanLibraryFunctionPointers();
-		FreeLibrary(vulkan_module);
-		vulkan_module = nullptr;
+		FreeLibrary(s_vulkan_module);
+		s_vulkan_module = nullptr;
 		return false;
 	}
 
-	vulkan_module_ref_count++;
 	return true;
 }
 
 void Vulkan::UnloadVulkanLibrary()
 {
-	if ((--vulkan_module_ref_count) > 0)
-		return;
-
 	ResetVulkanLibraryFunctionPointers();
-	FreeLibrary(vulkan_module);
-	vulkan_module = nullptr;
+	if (s_vulkan_module)
+		FreeLibrary(s_vulkan_module);
+	s_vulkan_module = nullptr;
 }
 
 #else
 
-static void* vulkan_module;
-static std::atomic_int vulkan_module_ref_count = {0};
+static void* s_vulkan_module;
+
+bool Vulkan::IsVulkanLibraryLoaded()
+{
+	return s_vulkan_module != nullptr;
+}
 
 bool Vulkan::LoadVulkanLibrary()
 {
-	// Not thread safe if a second thread calls the loader whilst the first is still in-progress.
-	if (vulkan_module)
-	{
-		vulkan_module_ref_count++;
-		return true;
-	}
+	pxAssertRel(!s_vulkan_module, "Vulkan module is not loaded.");
 
 #if defined(__APPLE__)
 	// Check if a path to a specific Vulkan library has been specified.
 	char* libvulkan_env = getenv("LIBVULKAN_PATH");
 	if (libvulkan_env)
-		vulkan_module = dlopen(libvulkan_env, RTLD_NOW);
-	if (!vulkan_module)
+		s_vulkan_module = dlopen(libvulkan_env, RTLD_NOW);
+	if (!s_vulkan_module)
 	{
 		unsigned path_size = 0;
 		_NSGetExecutablePath(nullptr, &path_size);
@@ -143,24 +139,24 @@ bool Vulkan::LoadVulkanLibrary()
 			{
 				path.erase(pos);
 				path += "/../Frameworks/libMoltenVK.dylib";
-				vulkan_module = dlopen(path.c_str(), RTLD_NOW);
+				s_vulkan_module = dlopen(path.c_str(), RTLD_NOW);
 			}
 		}
 	}
-	if (!vulkan_module)
-		vulkan_module = dlopen("libvulkan.dylib", RTLD_NOW);
+	if (!s_vulkan_module)
+		s_vulkan_module = dlopen("libvulkan.dylib", RTLD_NOW);
 #else
 	// Names of libraries to search. Desktop should use libvulkan.so.1 or libvulkan.so.
 	static const char* search_lib_names[] = {"libvulkan.so.1", "libvulkan.so"};
 	for (size_t i = 0; i < sizeof(search_lib_names) / sizeof(search_lib_names[0]); i++)
 	{
-		vulkan_module = dlopen(search_lib_names[i], RTLD_NOW);
-		if (vulkan_module)
+		s_vulkan_module = dlopen(search_lib_names[i], RTLD_NOW);
+		if (s_vulkan_module)
 			break;
 	}
 #endif
 
-	if (!vulkan_module)
+	if (!s_vulkan_module)
 	{
 		std::fprintf(stderr, "Failed to load or locate libvulkan.so\n");
 		return false;
@@ -168,7 +164,7 @@ bool Vulkan::LoadVulkanLibrary()
 
 	bool required_functions_missing = false;
 	auto LoadFunction = [&](void** func_ptr, const char* name, bool is_required) {
-		*func_ptr = dlsym(vulkan_module, name);
+		*func_ptr = dlsym(s_vulkan_module, name);
 		if (!(*func_ptr) && is_required)
 		{
 			std::fprintf(stderr, "Vulkan: Failed to load required module function %s\n", name);
@@ -183,23 +179,20 @@ bool Vulkan::LoadVulkanLibrary()
 	if (required_functions_missing)
 	{
 		ResetVulkanLibraryFunctionPointers();
-		dlclose(vulkan_module);
-		vulkan_module = nullptr;
+		dlclose(s_vulkan_module);
+		s_vulkan_module = nullptr;
 		return false;
 	}
 
-	vulkan_module_ref_count++;
 	return true;
 }
 
 void Vulkan::UnloadVulkanLibrary()
 {
-	if ((--vulkan_module_ref_count) > 0)
-		return;
-
 	ResetVulkanLibraryFunctionPointers();
-	dlclose(vulkan_module);
-	vulkan_module = nullptr;
+	if (s_vulkan_module)
+		dlclose(s_vulkan_module);
+	s_vulkan_module = nullptr;
 }
 
 #endif
