+++ pcsx2/GS/Renderers/Vulkan/VKStreamBuffer.cpp
@@ -15,9 +15,9 @@
 
 #include "PrecompiledHeader.h"
 
+#include "GS/Renderers/Vulkan/GSDeviceVK.h"
 #include "GS/Renderers/Vulkan/VKStreamBuffer.h"
-#include "GS/Renderers/Vulkan/VKContext.h"
-#include "GS/Renderers/Vulkan/VKUtil.h"
+#include "GS/Renderers/Vulkan/VKBuilders.h"
 
 #include "common/Align.h"
 #include "common/Assertions.h"
@@ -79,7 +79,7 @@ bool VKStreamBuffer::Create(VkBufferUsageFlags usage, u32 size)
 	VmaAllocationInfo ai = {};
 	VkBuffer new_buffer = VK_NULL_HANDLE;
 	VmaAllocation new_allocation = VK_NULL_HANDLE;
-	VkResult res = vmaCreateBuffer(g_vulkan_context->GetAllocator(), &bci, &aci, &new_buffer, &new_allocation, &ai);
+	VkResult res = vmaCreateBuffer(GSDeviceVK::GetInstance()->GetAllocator(), &bci, &aci, &new_buffer, &new_allocation, &ai);
 	if (res != VK_SUCCESS)
 	{
 		LOG_VULKAN_ERROR(res, "vkCreateBuffer failed: ");
@@ -105,9 +105,9 @@ void VKStreamBuffer::Destroy(bool defer)
 	if (m_buffer != VK_NULL_HANDLE)
 	{
 		if (defer)
-			g_vulkan_context->DeferBufferDestruction(m_buffer, m_allocation);
+			GSDeviceVK::GetInstance()->DeferBufferDestruction(m_buffer, m_allocation);
 		else
-			vmaDestroyBuffer(g_vulkan_context->GetAllocator(), m_buffer, m_allocation);
+			vmaDestroyBuffer(GSDeviceVK::GetInstance()->GetAllocator(), m_buffer, m_allocation);
 	}
 
 	m_size = 0;
@@ -194,7 +194,7 @@ void VKStreamBuffer::CommitMemory(u32 final_num_bytes)
 	pxAssert(final_num_bytes <= m_current_space);
 
 	// For non-coherent mappings, flush the memory range
-	vmaFlushAllocation(g_vulkan_context->GetAllocator(), m_allocation, m_current_offset, final_num_bytes);
+	vmaFlushAllocation(GSDeviceVK::GetInstance()->GetAllocator(), m_allocation, m_current_offset, final_num_bytes);
 
 	m_current_offset += final_num_bytes;
 	m_current_space -= final_num_bytes;
@@ -204,7 +204,7 @@ void VKStreamBuffer::CommitMemory(u32 final_num_bytes)
 void VKStreamBuffer::UpdateCurrentFencePosition()
 {
 	// Has the offset changed since the last fence?
-	const u64 counter = g_vulkan_context->GetCurrentFenceCounter();
+	const u64 counter = GSDeviceVK::GetInstance()->GetCurrentFenceCounter();
 	if (!m_tracked_fences.empty() && m_tracked_fences.back().first == counter)
 	{
 		// Still haven't executed a command buffer, so just update the offset.
@@ -221,7 +221,7 @@ void VKStreamBuffer::UpdateGPUPosition()
 	auto start = m_tracked_fences.begin();
 	auto end = start;
 
-	const u64 completed_counter = g_vulkan_context->GetCompletedFenceCounter();
+	const u64 completed_counter = GSDeviceVK::GetInstance()->GetCompletedFenceCounter();
 	while (end != m_tracked_fences.end() && completed_counter >= end->first)
 	{
 		m_current_gpu_position = end->second;
@@ -308,11 +308,11 @@ bool VKStreamBuffer::WaitForClearSpace(u32 num_bytes)
 
 	// Did any fences satisfy this condition?
 	// Has the command buffer been executed yet? If not, the caller should execute it.
-	if (iter == m_tracked_fences.end() || iter->first == g_vulkan_context->GetCurrentFenceCounter())
+	if (iter == m_tracked_fences.end() || iter->first == GSDeviceVK::GetInstance()->GetCurrentFenceCounter())
 		return false;
 
 	// Wait until this fence is signaled. This will fire the callback, updating the GPU position.
-	g_vulkan_context->WaitForFenceCounter(iter->first);
+	GSDeviceVK::GetInstance()->WaitForFenceCounter(iter->first);
 	m_tracked_fences.erase(
 		m_tracked_fences.begin(), m_current_offset == iter->second ? m_tracked_fences.end() : ++iter);
 	m_current_offset = new_offset;
