+++ pcsx2/pcsx2.vcxproj.filters
@@ -1310,9 +1310,6 @@
     <ClCompile Include="GS\Renderers\Vulkan\VKBuilders.cpp">
       <Filter>System\Ps2\GS\Renderers\Vulkan</Filter>
     </ClCompile>
-    <ClCompile Include="GS\Renderers\Vulkan\VKContext.cpp">
-      <Filter>System\Ps2\GS\Renderers\Vulkan</Filter>
-    </ClCompile>
     <ClCompile Include="GS\Renderers\Vulkan\VKLoader.cpp">
       <Filter>System\Ps2\GS\Renderers\Vulkan</Filter>
     </ClCompile>
@@ -1323,9 +1320,6 @@
       <Filter>System\Ps2\GS\Renderers\Vulkan</Filter>
     </ClCompile>
     <ClCompile Include="GS\Renderers\Vulkan\VKSwapChain.cpp">
-      <Filter>System\Ps2\GS\Renderers\Vulkan</Filter>
-    </ClCompile>
-    <ClCompile Include="GS\Renderers\Vulkan\VKUtil.cpp">
       <Filter>System\Ps2\GS\Renderers\Vulkan</Filter>
     </ClCompile>
     <ClCompile Include="Dmac.cpp">
@@ -2256,9 +2250,6 @@
     <ClInclude Include="GS\Renderers\Vulkan\VKBuilders.h">
       <Filter>System\Ps2\GS\Renderers\Vulkan</Filter>
     </ClInclude>
-    <ClInclude Include="GS\Renderers\Vulkan\VKContext.h">
-      <Filter>System\Ps2\GS\Renderers\Vulkan</Filter>
-    </ClInclude>
     <ClInclude Include="GS\Renderers\Vulkan\VKEntryPoints.h">
       <Filter>System\Ps2\GS\Renderers\Vulkan</Filter>
     </ClInclude>
@@ -2272,9 +2263,6 @@
       <Filter>System\Ps2\GS\Renderers\Vulkan</Filter>
     </ClInclude>
     <ClInclude Include="GS\Renderers\Vulkan\VKSwapChain.h">
-      <Filter>System\Ps2\GS\Renderers\Vulkan</Filter>
-    </ClInclude>
-    <ClInclude Include="GS\Renderers\Vulkan\VKUtil.h">
       <Filter>System\Ps2\GS\Renderers\Vulkan</Filter>
     </ClInclude>
     <ClInclude Include="ps2\pgif.h">
